# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/records": {
      "post": {
        "operationId": "createRecord",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Record created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "summary": "Creates a new record",
        "tags": [
          "Record"
        ]
      },
      "get": {
        "operationId": "searchRecords",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "description": "The cursor to start from",
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "The number of items to return",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "description": "The fields to sort by",
            "schema": {
              "properties": {
                "createdAt": {
                  "type": "string",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                },
                "updatedAt": {
                  "type": "string",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                },
                "eventDate": {
                  "type": "string",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              },
              "nullable": true,
              "type": "object"
            }
          },
          {
            "name": "term",
            "required": true,
            "in": "query",
            "description": "Search term to filter records by description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "query",
            "description": "Filter by patient ID",
            "schema": {
              "format": "uuid",
              "example": "dc5782d4-ce8e-4f95-bb0b-d1a4e130aee8",
              "type": "string"
            }
          },
          {
            "name": "attendanceType",
            "required": true,
            "in": "query",
            "description": "Filter by attendance type",
            "schema": {
              "type": "string",
              "enum": [
                "FIRST_VISIT",
                "FOLLOW_UP",
                "EVALUATION",
                "PROCEDURE",
                "TELEMEDICINE",
                "INTERCURRENCE"
              ]
            }
          },
          {
            "name": "clinicalStatus",
            "required": true,
            "in": "query",
            "description": "Filter by clinical status",
            "schema": {
              "type": "string",
              "enum": [
                "STABLE",
                "IMPROVING",
                "WORSENING",
                "UNCHANGED",
                "UNDER_OBSERVATION"
              ]
            }
          },
          {
            "name": "dateStart",
            "required": true,
            "in": "query",
            "description": "Filter events from this date (eventDate)",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "dateEnd",
            "required": true,
            "in": "query",
            "description": "Filter events up to this date (eventDate)",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": true,
            "in": "query",
            "description": "Filter by record source",
            "schema": {
              "type": "string",
              "enum": [
                "MANUAL",
                "IMPORT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Records list"
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "summary": "Lists and searches records",
        "tags": [
          "Record"
        ]
      }
    },
    "/api/v1/records/{id}": {
      "get": {
        "operationId": "getRecord",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "summary": "Gets a record by ID",
        "tags": [
          "Record"
        ]
      },
      "put": {
        "operationId": "updateRecord",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Record updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "summary": "Updates a record",
        "tags": [
          "Record"
        ]
      },
      "delete": {
        "operationId": "deleteRecord",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Record ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record deleted"
          },
          "default": {
            "description": "Request failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "summary": "Deletes a record",
        "tags": [
          "Record"
        ]
      }
    },
...
  },
  "info": {
    "title": "Automo API",
    "description": "The backend API for the Automo project<h3>Authentication</h3>\n                For the protected routes, you must first <b>sign up</b> (<code>/user/sign-up</code>) and then <b>sign in</b> (<code>/auth/sign-in</code>).\n                After that, you already can request the protected routes.\n                The API will return a token that will be included automatically in the cookie and will be sent in subsequent requests for authentication.\n                To sign out, simply make a request to the <code>/auth/sign-out</code> endpoint,\n                the token will be invalidated and removed from the cookie.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
      "ApiProblem": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem",
            "example": "Invalid input"
          },
          "type": {
            "type": "object",
            "format": "uri",
            "description": "A URI reference that identifies the problem type",
            "example": "https://developer.mozilla.org/docs/Web/HTTP/Status/400"
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem",
            "example": "The ID is required"
          },
          "status": {
            "type": "number",
            "description": "The HTTP status code",
            "example": 400,
            "enum": [
              400,
              401,
              402,
              403,
              404,
              405,
              406,
              407,
              408,
              409,
              410,
              411,
              412,
              413,
              414,
              415,
              416,
              417,
              418,
              421,
              422,
              423,
              424,
              428,
              429,
              456,
              500,
              501,
              502,
              503,
              504,
              505,
              507,
              508
            ]
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/example/12345"
          }
        },
        "required": [
          "title",
          "type",
          "detail",
          "status",
          "instance"
        ]
      },
...
      "CreateRecordDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid",
            "example": "dc5782d4-ce8e-4f95-bb0b-d1a4e130aee8"
          },
          "professionalId": {
            "type": "string",
            "format": "uuid",
            "example": "dc5782d4-ce8e-4f95-bb0b-d1a4e130aee8"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "General clinical note"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fileName": {
                  "type": "string",
                  "minLength": 1,
                  "example": "exam.pdf"
                },
                "url": {
                  "type": "string",
                  "format": "uri",
                  "example": "https://storage.example.com/exam.pdf"
                },
                "description": {
                  "type": "string",
                  "example": "Blood test results"
                }
              },
              "required": [
                "fileName",
                "url",
                "description"
              ]
            },
            "default": []
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "example": "Follow-up consultation"
          },
          "templateType": {
            "type": "string",
            "enum": [
              "SOAP",
              "DAP"
            ]
          },
          "attendanceType": {
            "type": "string",
            "enum": [
              "FIRST_VISIT",
              "FOLLOW_UP",
              "EVALUATION",
              "PROCEDURE",
              "TELEMEDICINE",
              "INTERCURRENCE"
            ]
          },
          "clinicalStatus": {
            "type": "string",
            "enum": [
              "STABLE",
              "IMPROVING",
              "WORSENING",
              "UNCHANGED",
              "UNDER_OBSERVATION"
            ]
          },
          "conductTags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GUIDANCE",
                "PRESCRIPTION",
                "EXAM_REQUESTED",
                "REFERRAL",
                "FOLLOW_UP_SCHEDULED",
                "THERAPY_ADJUSTMENT"
              ]
            },
            "default": []
          },
          "subjective": {
            "type": "string",
            "example": "Patient reports persistent headache for 3 days"
          },
          "objective": {
            "type": "string",
            "example": "BP 120/80, no fever"
          },
          "assessment": {
            "type": "string",
            "example": "Tension headache, likely stress-related"
          },
          "plan": {
            "type": "string",
            "example": "Analgesics as needed, return in 7 days if unresolved"
          },
          "freeNotes": {
            "type": "string",
            "example": "Patient was anxious during the visit"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time",
            "description": "A datetime in ISO 8601 format."
          },
          "appointmentId": {
            "type": "string",
            "format": "uuid",
            "example": "dc5782d4-ce8e-4f95-bb0b-d1a4e130aee8"
          },
          "source": {
            "type": "string",
            "enum": [
              "MANUAL",
              "IMPORT"
            ]
          },
          "importedDocumentId": {
            "type": "string",
            "format": "uuid"
          },
          "wasHumanEdited": {
            "type": "boolean"
          }
        },
        "required": [
          "patientId",
          "professionalId"
        ]
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fileName",
          "url",
          "description"
        ]
      },
      "Record": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the entity"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time the entity was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time the entity was last updated"
          },
          "patientId": {
            "type": "string",
            "format": "uuid",
            "description": "The patient ID"
          },
          "professionalId": {
            "type": "string",
            "format": "uuid",
            "description": "The professional ID"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "description": "Legacy free-text description"
          },
          "templateType": {
            "type": "string",
            "enum": [
              "SOAP",
              "DAP"
            ],
            "nullable": true,
            "description": "Evolution template type"
          },
          "title": {
            "type": "object",
            "nullable": true,
            "description": "Evolution title"
          },
          "attendanceType": {
            "type": "string",
            "enum": [
              "FIRST_VISIT",
              "FOLLOW_UP",
              "EVALUATION",
              "PROCEDURE",
              "TELEMEDICINE",
              "INTERCURRENCE"
            ],
            "nullable": true,
            "description": "Attendance type"
          },
          "clinicalStatus": {
            "type": "string",
            "enum": [
              "STABLE",
              "IMPROVING",
              "WORSENING",
              "UNCHANGED",
              "UNDER_OBSERVATION"
            ],
            "nullable": true,
            "description": "Clinical status"
          },
          "conductTags": {
            "type": "array",
            "description": "Conduct tags",
            "items": {
              "type": "string",
              "enum": [
                "GUIDANCE",
                "PRESCRIPTION",
                "EXAM_REQUESTED",
                "REFERRAL",
                "FOLLOW_UP_SCHEDULED",
                "THERAPY_ADJUSTMENT"
              ]
            }
          },
          "subjective": {
            "type": "object",
            "nullable": true,
            "description": "SOAP Subjective"
          },
          "objective": {
            "type": "object",
            "nullable": true,
            "description": "SOAP Objective"
          },
          "assessment": {
            "type": "object",
            "nullable": true,
            "description": "SOAP/DAP Assessment"
          },
          "plan": {
            "type": "object",
            "nullable": true,
            "description": "SOAP/DAP Plan"
          },
          "freeNotes": {
            "type": "object",
            "nullable": true,
            "description": "Free notes"
          },
          "eventDate": {
            "type": "object",
            "format": "date-time",
            "nullable": true,
            "description": "Event date (clinical event)"
          },
          "appointmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "Linked appointment ID"
          },
          "source": {
            "type": "string",
            "enum": [
              "MANUAL",
              "IMPORT"
            ],
            "description": "Origin of the record"
          },
          "importedDocumentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "ID of the imported document if applicable"
          },
          "wasHumanEdited": {
            "type": "boolean",
            "description": "Indicates if the imported record was manually reviewed and edited"
          },
          "files": {
            "description": "The attached files",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDto"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "patientId",
          "professionalId",
          "description",
          "templateType",
          "title",
          "attendanceType",
          "clinicalStatus",
          "conductTags",
          "subjective",
          "objective",
          "assessment",
          "plan",
          "freeNotes",
          "eventDate",
          "appointmentId",
          "source",
          "importedDocumentId",
          "wasHumanEdited",
          "files"
        ]
      },
      "UpdateRecordInputDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "Updated clinical note"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "example": "Follow-up consultation"
          },
          "templateType": {
            "type": "string",
            "enum": [
              "SOAP",
              "DAP"
            ]
          },
          "attendanceType": {
            "type": "string",
            "enum": [
              "FIRST_VISIT",
              "FOLLOW_UP",
              "EVALUATION",
              "PROCEDURE",
              "TELEMEDICINE",
              "INTERCURRENCE"
            ]
          },
          "clinicalStatus": {
            "type": "string",
            "enum": [
              "STABLE",
              "IMPROVING",
              "WORSENING",
              "UNCHANGED",
              "UNDER_OBSERVATION"
            ]
          },
          "conductTags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "GUIDANCE",
                "PRESCRIPTION",
                "EXAM_REQUESTED",
                "REFERRAL",
                "FOLLOW_UP_SCHEDULED",
                "THERAPY_ADJUSTMENT"
              ]
            }
          },
          "subjective": {
            "type": "string"
          },
          "objective": {
            "type": "string"
          },
          "assessment": {
            "type": "string"
          },
          "plan": {
            "type": "string"
          },
          "freeNotes": {
            "type": "string"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time",
            "description": "A datetime in ISO 8601 format."
          },
          "source": {
            "type": "string",
            "enum": [
              "MANUAL",
              "IMPORT"
            ]
          },
          "importedDocumentId": {
            "type": "string",
            "format": "uuid"
          },
          "wasHumanEdited": {
            "type": "boolean"
          }
        },
        "required": [
          "description",
          "title",
          "templateType",
          "attendanceType",
          "clinicalStatus",
          "conductTags",
          "subjective",
          "objective",
          "assessment",
          "plan",
          "freeNotes",
          "eventDate",
          "source",
          "importedDocumentId",
          "wasHumanEdited"
        ]
      },
...
    }
  }
}